Feature/57-Create-a-calendar-page (#110)
* Code of Conduct

* Delete CODE_OF_CONDUCT

* CODE_OF_CONDUCT

The following document specifies the etiquette that should be following when using Taffy's OSS and also the following:
Where the code of conduct takes effect?
Whom the code of conduct applies to?
What happens if someone violates the code of conduct?
How someone can report violations?

* Delete CODE OF CONDUCT

* Removed repeated Calendar page

* Removing unused repeated files

* Calendar Component with List View

* Fixed Broken List View

* Rendering the calendar only once
