--- doto-frontend/src/components/pages/Calendar.js
@@ -1,13 +1,0 @@
-import React from "react";
-
-const Calendar = () => {
-    return (
-        <div className="App">
-            <header className="App-header">
-                <div>Calendar</div>
-            </header>
-        </div>
-    );
-};
-
-export default Calendar;
